+++ plugins/plugin-ui/src/components/shape/index.ts
@@ -1,6 +1,6 @@
 export { AnchoredPanelPortal } from "./AnchoredPanelPortal";
 export type { AnchoredPanelPortalProps } from "./AnchoredPanelPortal";
-export { resolveDelpiUiPortalTheme, DELPI_UI_SHAPE_THEME_HOST_CLASS } from "./delpiUiPortalTheme";
+export { resolveDelpiUiPortalTheme, resolveMfePortalScopeClassName, DELPI_UI_SHAPE_THEME_HOST_CLASS } from "./delpiUiPortalTheme";
 export { useDelpiUiPortalTheme } from "./useDelpiUiPortalTheme";
 export { ColorThemeGrid, ColorStandardRow, ColorSwatch } from "./ColorThemeGrid";
 export { ColorPickerPopover, ColorPickerPopoverTrigger, ShapeFillMenu } from "./ColorPickerPopover";